metaagent/tools/fsgan/requirements.txt
setuptools==58.2.0
torch-summary==1.4.5
opencv-contrib-python==4.5.4.60
tensorflow==2.7.0
tqdm==4.64.1
matplotlib==3.6.2
ffmpeg-python==0.2.0
PyYAML==6.0
pandas==1.5.1
seaborn==0.12.1
scipy==1.9.3